+++ symgene/combiners/__init__.py
@@ -1,1 +1,6 @@
-﻿
+from symgene.combiners.base import BaseCombiner
+from symgene.combiners.ridge import RidgeCombiner
+from symgene.combiners.lasso import LassoCombiner
+from symgene.combiners.linear import LinearCombiner
+
+__all__ = ["BaseCombiner", "RidgeCombiner", "LassoCombiner", "LinearCombiner"]